FWA collection

Get and filter collection from database.

fwa_collection(
  table,
  filter = NULL,
  bbox = NULL,
  columns = NULL,
  epsg = 4326,
  limit = 10000,
  offset = 0,
  schema = "whse_basemapping"
)

Arguments

table
A character string of the table name.

filter
A named vector or list of the filter(s) to apply, where the list names correspond to column names and the list values correspond to the desired value, e.g. list(gnis_name = "Sangan River").

bbox
A vector of numbers indicating bounding box to limit output features to, e.g. c(minLon, minLat, maxLon, maxLat).

columns
A vector of strings of the column names to include. If NULL (default), all columns are retained.

epsg
A positive whole number of the epsg to transform features to.

limit
A positive whole number indicating the maximum number of features to return.

offset
A positive whole number indicating the offset of start of returned results.

schema
A character string of the schema to prepend to collection_id. Possible values include 'whse_basemapping', 'usgs', 'hydrosheds' and NULL. If NULL, nothing is prepended.

Value

A sf object

Examples

fwa_collection("fwa_stream_networks_sp", filter = list(gnis_name = 'Sangan River'))

#> Simple feature collection with 21 features and 28 fields
#> geometry type:  LINESTRING
#> dimension:      XYZ
#> bbox:           xmin: -131.9972 ymin: 53.94994 xmax: -131.9168 ymax: 54.0332
#> z_range:        zmin: 1 zmax: 67
#> geographic CRS: WGS 84
#> First 10 features:
#>          id blue_line_key blue_line_key_50k downstream_route_measure edge_type
#> 1  59028661     360879896                11                4720.9794      1000
#> 2  59028739     360879896                11                3446.3450      1000
#> 3  59028963     360879896                11                2318.8188      1000
#> 4  59029083     360879896                11                   0.0000      1250
#> 5  59029299     360879896                11                1266.2343      1000
#> 6  59029455     360879896                11                 366.7395      1250
#> 7  59031813     360879896                11                5028.8671      1000
#> 8  59032427     360879896                11                6466.4298      1000
#> 9  59033632     360879896                11                6807.7279      1000
#> 10 59034768     360879896                11                7384.3944      1000
#>    feature_code  feature_source
#> 1    GA24850000 linear stream -
#> 2    GA24850000 linear stream -
#> 3    GA24850000 linear stream -
#> 4    WA24111120 areal stream sk
#> 5    GA24850000 linear stream -
#> 6    WA24111120 areal stream sk
#> 7    GA24850000 linear stream -
#> 8    GA24850000 linear stream -
#> 9    GA24850000 linear stream -
#> 10   GA24850000 linear stream -
#>                                                                                                                                 fwa_watershed_code
#> 1  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 2  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 3  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 4  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 5  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 6  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 7  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 8  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 9  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 10 940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#>    gnis_id    gnis_name gradient left_right_tributary length_metre
#> 1     3390 Sangan River   0.0032                 NONE     307.8877
#> 2     3390 Sangan River   0.0000                 NONE    1274.6344
#> 3     3390 Sangan River   0.0009                 NONE    1127.5262
#> 4     3390 Sangan River   0.0136                 NONE     366.7395
#> 5     3390 Sangan River   0.0000                 NONE    1052.5845
#> 6     3390 Sangan River   0.0000                 NONE     899.4948
#> 7     3390 Sangan River   0.0146                 NONE    1437.5627
#> 8     3390 Sangan River   0.0059                 NONE     341.2980
#> 9     3390 Sangan River   0.0069                 NONE     576.6665
#> 10    3390 Sangan River   0.0108                 NONE     464.3463
#>    linear_feature_id
#> 1           59028661
#> 2           59028739
#> 3           59028963
#> 4           59029083
#> 5           59029299
#> 6           59029455
#> 7           59031813
#> 8           59032427
#> 9           59033632
#> 10          59034768
#>                                                                                                                               local_watershed_code
#> 1  940-971473-217852-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 2  940-971473-159033-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 3  940-971473-107003-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 4  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 5  940-971473-016923-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 6  940-971473-016923-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 7  940-971473-232060-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 8  940-971473-298397-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 9  940-971473-314146-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 10 940-971473-340757-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#>      localcode_ltree stream_magnitude stream_order upstream_area_ha
#> 1  940.971473.217852               46            4         5063.281
#> 2  940.971473.159033               47            4         5207.201
#> 3  940.971473.107003               60            4         5823.088
#> 4         940.971473              238            5        15398.811
#> 5  940.971473.016923              230            5        14384.285
#> 6  940.971473.016923              230            5        14384.285
#> 7  940.971473.232060               45            4         4931.980
#> 8  940.971473.298397               44            4         4836.334
#> 9  940.971473.314146               33            3         4108.151
#> 10 940.971473.340757               32            3         4056.546
#>    upstream_route_measure waterbody_key
#> 1               5028.8673            NA
#> 2               4720.9800            NA
#> 3               3446.3452            NA
#> 4                366.7394     329164263
#> 5               2318.8184            NA
#> 6               1266.2350     329164263
#> 7               6466.4297            NA
#> 8               6807.7272            NA
#> 9               7384.3934            NA
#> 10              7848.7408            NA
#>                               watershed_code_50k watershed_group_code
#> 1  940974000000000000000000000000000000000000000                 GRAI
#> 2  940974000000000000000000000000000000000000000                 GRAI
#> 3  940974000000000000000000000000000000000000000                 GRAI
#> 4  940974000000000000000000000000000000000000000                 GRAI
#> 5  940974000000000000000000000000000000000000000                 GRAI
#> 6  940974000000000000000000000000000000000000000                 GRAI
#> 7  940974000000000000000000000000000000000000000                 GRAI
#> 8  940974000000000000000000000000000000000000000                 GRAI
#> 9  940974000000000000000000000000000000000000000                 GRAI
#> 10 940974000000000000000000000000000000000000000                 GRAI
#>    watershed_group_code_50k watershed_group_id watershed_key watershed_key_50k
#> 1                      GRAI                 60     360879896                11
#> 2                      GRAI                 60     360879896                11
#> 3                      GRAI                 60     360879896                11
#> 4                      GRAI                 60     360879896                11
#> 5                      GRAI                 60     360879896                11
#> 6                      GRAI                 60     360879896                11
#> 7                      GRAI                 60     360879896                11
#> 8                      GRAI                 60     360879896                11
#> 9                      GRAI                 60     360879896                11
#> 10                     GRAI                 60     360879896                11
#>    wscode_ltree                       geometry
#> 1    940.971473 LINESTRING Z (-131.9285 54....
#> 2    940.971473 LINESTRING Z (-131.9463 54....
#> 3    940.971473 LINESTRING Z (-131.9631 54....
#> 4    940.971473 LINESTRING Z (-131.9972 54....
#> 5    940.971473 LINESTRING Z (-131.9786 54....
#> 6    940.971473 LINESTRING Z (-131.9919 54....
#> 7    940.971473 LINESTRING Z (-131.9273 54....
#> 8    940.971473 LINESTRING Z (-131.9196 54....
#> 9    940.971473 LINESTRING Z (-131.9195 54....
#> 10   940.971473 LINESTRING Z (-131.9172 54....