Freshwater Atlas Locate Along Interval

Gets a regularly spaced set of points between two river metres.

fwa_locate_along_interval(
  blue_line_key,
  interval_length = 1000,
  start_measure = 0,
  end_measure = NULL,
  limit = 100,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)

Arguments

blue_line_key
A positive whole number of the stream blue line key.

interval_length
An integer of the interval distance in meters.

start_measure
An integer of the distance in meters upstream from the river mouth to start from.

end_measure
An integer of the distance in meters upstream from the river mouth to end at.

limit
A positive whole number indicating the maximum number of features to return.

offset
A positive whole number indicating the offset of start of returned results.

bbox
A vector of four numbers indicating bounding box to limit output features to, e.g. c(minLon, minLat, maxLon, maxLat).

properties
A vector of strings of the column names to include. If NULL (default), all columns are retained.

transform
A character vector with the name of the transform function followed by the parameter values (e.g. c("ST_Simplify", 100))

epsg
A positive whole number of the epsg to transform features to.

Value

An sf object.

See also

Examples

fwa_locate_along_interval(356308001, interval_length = 10, start_measure = 0)
#> Simple feature collection with 100 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XYZ
#> Bounding box:  xmin: -122.0155 ymin: 49.16486 xmax: -122.0047 ymax: 49.16804
#> z_range:       zmin: 9.626 zmax: 10.131
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    downstream_route_measure index                       geometry
#> 1                         0     0 POINT Z (-122.0155 49.16804...
#> 2                        10     1 POINT Z (-122.0154 49.16795...
#> 3                        20     2 POINT Z (-122.0154 49.16787...
#> 4                        30     3 POINT Z (-122.0153 49.16778...
#> 5                        40     4 POINT Z (-122.0153 49.1677 ...
#> 6                        50     5 POINT Z (-122.0152 49.16762...
#> 7                        60     6 POINT Z (-122.0152 49.16753...
#> 8                        70     7 POINT Z (-122.0151 49.16745...
#> 9                        80     8 POINT Z (-122.0151 49.16736...
#> 10                       90     9 POINT Z (-122.015 49.16728 ...