FWA Watershed at Downstream Route Measure

Gets a polygon of the watershed at a specified river metre.

fwa_watershed_at_measure(
  blue_line_key,
  downstream_route_measure = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)

Arguments

blue_line_key
A positive whole number of the stream blue line key.

downstream_route_measure
A positive number of the river metre.

bbox
A vector of four numbers indicating bounding box to limit output features to, e.g. c(minLon, minLat, maxLon, maxLat).

properties
A vector of strings of the column names to include. If NULL (default), all columns are retained.

transform
A character vector with the name of the transform function followed by the parameter values (e.g. c("ST_Simplify", 100))

epsg
A positive whole number of the epsg to transform features to.

Value

An sf object.

See also

Examples

fwa_watershed_at_measure(356308001, downstream_route_measure = 10000)
#> Simple feature collection with 1 feature and 4 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -121.9625 ymin: 49.09301 xmax: -121.8528 ymax: 49.14291
#> Geodetic CRS:  WGS 84
#>   area_ha   localcode_ltree refine_method wscode_ltree
#> 1 2364.24 100.070889.490268           DEM   100.070889
#>                         geometry
#> 1 POLYGON ((-121.9625 49.1038...