FWA collection

Get and filter collection from database.

fwa_collection(
  collection_id,
  filter = NULL,
  limit = 100,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)

Arguments

collection_id
A character string of the collection id.

filter
A named vector or list of the filter(s) to apply, where the list names correspond to column names and the list values correspond to the desired value, e.g. list(gnis_name = "Sangan River").

limit
A positive whole number indicating the maximum number of features to return.

offset
A positive whole number indicating the offset of start of returned results.

bbox
A vector of four numbers indicating bounding box to limit output features to, e.g. c(minLon, minLat, maxLon, maxLat).

properties
A vector of strings of the column names to include. If NULL (default), all columns are retained.

transform
A character vector with the name of the transform function followed by the parameter values (e.g. c("ST_Simplify", 100))

epsg
A positive whole number of the epsg to transform features to.

Value

An sf object.

See also

Examples

collection_id <- "whse_basemapping.fwa_stream_networks_sp"
fwa_collection(collection_id, filter = list(gnis_name = "Sangan River"))
#> Simple feature collection with 21 features and 28 fields
#> Geometry type: LINESTRING
#> Dimension:     XYZ
#> Bounding box:  xmin: -131.9972 ymin: 53.94994 xmax: -131.9168 ymax: 54.0332
#> z_range:       zmin: 1 zmax: 67
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>          id blue_line_key blue_line_key_50k downstream_route_measure edge_type
#> 1  59029299     360879896                11                 1266.234      1000
#> 2  59038020     360879896                11                 8743.778      1000
#> 3  59028963     360879896                11                 2318.819      1000
#> 4  59029083     360879896                11                    0.000      1250
#> 5  59052964     360879896               130                17516.161      1050
#> 6  59040619     360879896                11                10075.283      1000
#> 7  59033632     360879896                11                 6807.728      1000
#> 8  59028661     360879896                11                 4720.979      1000
#> 9  59045671     360879896                11                13067.935      1000
#> 10 59036446     360879896                11                 7848.741      1000
#>    feature_code  feature_source
#> 1    GA24850000 linear stream -
#> 2    GA24850000 linear stream -
#> 3    GA24850000 linear stream -
#> 4    WA24111120 areal stream sk
#> 5    GA24850140 linear stream -
#> 6    GA24850000 linear stream -
#> 7    GA24850000 linear stream -
#> 8    GA24850000 linear stream -
#> 9    GA24850000 linear stream -
#> 10   GA24850000 linear stream -
#>                                                                                                                                 fwa_watershed_code
#> 1  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 2  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 3  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 4  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 5  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 6  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 7  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 8  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 9  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 10 940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#>    gnis_id    gnis_name gradient left_right_tributary length_metre
#> 1     3390 Sangan River   0.0000                 NONE    1052.5845
#> 2     3390 Sangan River   0.0038                 NONE     799.8396
#> 3     3390 Sangan River   0.0009                 NONE    1127.5262
#> 4     3390 Sangan River   0.0136                 NONE     366.7395
#> 5     3390 Sangan River   0.0000                 NONE      55.2556
#> 6     3390 Sangan River   0.0024                 NONE     833.9275
#> 7     3390 Sangan River   0.0069                 NONE     576.6665
#> 8     3390 Sangan River   0.0032                 NONE     307.8877
#> 9     3390 Sangan River   0.0042                 NONE     707.9944
#> 10    3390 Sangan River   0.0022                 NONE     895.0372
#>    linear_feature_id
#> 1           59029299
#> 2           59038020
#> 3           59028963
#> 4           59029083
#> 5           59052964
#> 6           59040619
#> 7           59033632
#> 8           59028661
#> 9           59045671
#> 10          59036446
#>                                                                                                                               local_watershed_code
#> 1  940-971473-016923-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 2  940-971473-403487-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 3  940-971473-107003-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 4  940-971473-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 5  940-971473-765465-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 6  940-971473-464930-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 7  940-971473-314146-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 8  940-971473-217852-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 9  940-971473-603028-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#> 10 940-971473-362185-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000-000000
#>      localcode_ltree stream_magnitude stream_order upstream_area_ha
#> 1  940.971473.016923              230            5             <NA>
#> 2  940.971473.403487               30            3             <NA>
#> 3  940.971473.107003               60            4             <NA>
#> 4         940.971473              238            5             <NA>
#> 5  940.971473.765465                4            2             <NA>
#> 6  940.971473.464930               28            3             <NA>
#> 7  940.971473.314146               33            3             <NA>
#> 8  940.971473.217852               46            4             <NA>
#> 9  940.971473.603028               20            3             <NA>
#> 10 940.971473.362185               31            3             <NA>
#>    upstream_route_measure waterbody_key
#> 1               2318.8184            NA
#> 2               9543.6177            NA
#> 3               3446.3452            NA
#> 4                366.7394     329164263
#> 5              17571.4170     329163195
#> 6              10909.2099            NA
#> 7               7384.3934            NA
#> 8               5028.8673            NA
#> 9              13775.9302            NA
#> 10              8743.7778            NA
#>                               watershed_code_50k watershed_group_code
#> 1  940974000000000000000000000000000000000000000                 GRAI
#> 2  940974000000000000000000000000000000000000000                 GRAI
#> 3  940974000000000000000000000000000000000000000                 GRAI
#> 4  940974000000000000000000000000000000000000000                 GRAI
#> 5  940974000880000000000000000000000000000000000                 GRAI
#> 6  940974000000000000000000000000000000000000000                 GRAI
#> 7  940974000000000000000000000000000000000000000                 GRAI
#> 8  940974000000000000000000000000000000000000000                 GRAI
#> 9  940974000000000000000000000000000000000000000                 GRAI
#> 10 940974000000000000000000000000000000000000000                 GRAI
#>    watershed_group_code_50k watershed_group_id watershed_key watershed_key_50k
#> 1                      GRAI                 60     360879896                11
#> 2                      GRAI                 60     360879896                11
#> 3                      GRAI                 60     360879896                11
#> 4                      GRAI                 60     360879896                11
#> 5                      GRAI                 60     360879896               130
#> 6                      GRAI                 60     360879896                11
#> 7                      GRAI                 60     360879896                11
#> 8                      GRAI                 60     360879896                11
#> 9                      GRAI                 60     360879896                11
#> 10                     GRAI                 60     360879896                11
#>    wscode_ltree                       geometry
#> 1    940.971473 LINESTRING Z (-131.9786 54....
#> 2    940.971473 LINESTRING Z (-131.9212 54....
#> 3    940.971473 LINESTRING Z (-131.9631 54....
#> 4    940.971473 LINESTRING Z (-131.9972 54....
#> 5    940.971473 LINESTRING Z (-131.9458 53....
#> 6    940.971473 LINESTRING Z (-131.9191 53....
#> 7    940.971473 LINESTRING Z (-131.9195 54....
#> 8    940.971473 LINESTRING Z (-131.9285 54....
#> 9    940.971473 LINESTRING Z (-131.9272 53....
#> 10   940.971473 LINESTRING Z (-131.9169 54....