Query Freshwater Atlas Collection

Query a collection from the freshwater atlas.

fwa_collection(
  collection_id,
  filter = NULL,
  limit = 10000,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)

fwa_query_collection(
  collection_id,
  filter = NULL,
  limit = 10000,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)

Arguments

collection_id

A character string of the collection id.

filter

A named vector or list of the filter(s) to apply, where the list names correspond to column names and the list values correspond to the desired value, e.g. list(gnis_name = "Sangan River").

limit

A positive whole number indicating the maximum number of features to return.

bbox

A vector of four numbers indicating bounding box to limit output features to, e.g. c(minLon, minLat, maxLon, maxLat).

properties

A vector of strings of the column names to include. If NULL (default), all columns are retained.

transform

A character vector with the name of the valid transform function followed by the parameter values (e.g. c("ST_Simplify", 100))

epsg

A positive whole number of the epsg to transform features to.

offset

A positive whole number indicating the offset of start of returned results.

Value

An sf object.

Functions

fwa_collection: Query Freshwater Atlas Collection

See also

Examples

collection_id <- "whse_basemapping.fwa_stream_networks_sp"
filter <- list(gnis_name = "Sangan River")
fwa_query_collection(collection_id, filter = filter)
#> Simple feature collection with 21 features and 28 fields
#> Geometry type: LINESTRING
#> Dimension:     XYZ
#> Bounding box:  xmin: -131.9972 ymin: 53.94994 xmax: -131.9168 ymax: 54.0332
#> z_range:       zmin: 1 zmax: 67
#> Geodetic CRS:  WGS 84
#> # A data frame: 21 × 29
#>    id     blue_line_key blue_line_key_5… downstream_rout… edge_type feature_code
#>    <chr>          <int>            <int>            <dbl>     <int> <chr>       
#>  1 59028…     360879896               11            4721.      1000 GA24850000  
#>  2 59028…     360879896               11            3446.      1000 GA24850000  
#>  3 59028…     360879896               11            2319.      1000 GA24850000  
#>  4 59029…     360879896               11               0       1250 WA24111120  
#>  5 59029…     360879896               11            1266.      1000 GA24850000  
#>  6 59029…     360879896               11             367.      1250 WA24111120  
#>  7 59031…     360879896               11            5029.      1000 GA24850000  
#>  8 59032…     360879896               11            6466.      1000 GA24850000  
#>  9 59033…     360879896               11            6808.      1000 GA24850000  
#> 10 59034…     360879896               11            7384.      1000 GA24850000  
#> # … with 11 more rows, and 23 more variables: feature_source <chr>,
#> #   fwa_watershed_code <chr>, gnis_id <int>, gnis_name <chr>, gradient <dbl>,
#> #   left_right_tributary <chr>, length_metre <dbl>, linear_feature_id <int>,
#> #   local_watershed_code <chr>, localcode_ltree <chr>, stream_magnitude <int>,
#> #   stream_order <int>, upstream_area_ha <chr>, upstream_route_measure <dbl>,
#> #   waterbody_key <int>, watershed_code_50k <chr>, watershed_group_code <chr>,
#> #   watershed_group_code_50k <chr>, watershed_group_id <int>, …