FWA watershed at downstream route measure

Provided a location as blue_line_key and downstream_route_measure, return the entire watershed boundary upstream of the location.

fwa_watershed_at_measure(
  blue_line_key,
  downstream_route_measure = 0,
  limit = 100,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  epsg = 4326
)

Arguments

blue_line_key
A positive whole number of the stream blue line key.

downstream_route_measure
A positive number of the downstream route measure.

limit
A positive whole number indicating the maximum number of features to return.

offset
A positive whole number indicating the offset of start of returned results.

bbox
A vector of numbers indicating bounding box to limit output features to, e.g. c(minLon, minLat, maxLon, maxLat).

properties
A vector of strings of the column names to include. If NULL (default), all columns are retained.

transform
A vector of strings of the geometry transform function followed by parameter values (e.g. c("ST_Simplify", 100))

epsg
A positive whole number of the epsg to transform features to.

Value

A sf object

See also

Examples

if (FALSE) fwa_watershed_at_measure(356308001, downstream_route_measure = 10000)